+++ src/services/urls.service.js
@@ -30,10 +30,12 @@ const getUrlById = async (id) => {
   }
 };
 
-const deleteUrlById = async (id) => {
+const deleteUrlById = async (id, userId) => {
   try {
     const result = await urlsRepository.getUrlById(id);
     if (result.rowCount === 0) return { status: 404, response: { message: "Url não encontrada" } };
+
+    if (result.rows[0].userId !== userId) return { status: 401, response: { message: "Você não tem permissão para realizar essa operação" } };
 
     await urlsRepository.deleteUrlById(id);
     return { status: 204, response: { message: "Url excluída com sucesso" } };
